Split node-specific compiler tests
